skip last finished event in cleanup

+++ services/transmitter/transmitter/logic.py
@@ -229,7 +229,16 @@ class EventProcessor:
             )
 
     def cleanup_sync_comment_events(self):
-        self._log.info("Cleaning up comment sync events.")
+        self._log.debug("Cleaning up comment sync events.")
+        any_in_progress = self._cleanup_in_progress_comment_events()
+        if any_in_progress:
+            return
+
+        last_finished_event = self._get_last_finished_event()
+        last_event_id = None
+        if last_finished_event:
+            last_event_id = last_finished_event["id"]
+
         cleanup_date = arrow.utcnow() - datetime.timedelta(
             days=COMMENT_EVENTS_CLEANUP_TIMEOUT_DAYS
         )
@@ -240,8 +249,10 @@ class EventProcessor:
         ))
         removed = 0
         for event in events_to_cleanup:
+            event_id = event["id"]
+            if last_event_id == event_id:
+                continue
             try:
-                event_id = event["id"]
                 ayon_api.delete_event(event_id)
                 removed += 1
             except Exception:
@@ -249,7 +260,7 @@ class EventProcessor:
                     f"Failed to delete event {event_id}.",
                     exc_info=True
                 )
-        self._log.info(f"Cleaned up {removed} events.")
+        self._log.debug(f"Cleaned up {removed} events.")
 
     def _cleanup_in_progress_comment_events(self) -> bool:
         in_progress_events = list(ayon_api.get_events(
